Working evaluation scripts

+++ utils/dist_helper.py
@@ -12,10 +12,10 @@ from scipy.linalg import toeplitz
 
 def emd(x, y, distance_scaling=1.0):
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     support_size = max(len(x), len(y))
-    d_mat = toeplitz(range(support_size)).astype(np.float)  # diagonal-constant matrix
+    d_mat = toeplitz(range(support_size)).astype(float)  # diagonal-constant matrix
     distance_mat = d_mat / distance_scaling
     x, y = process_tensor(x, y)
 
@@ -39,8 +39,8 @@ def gaussian_emd(x, y, sigma=1.0, distance_scaling=1.0):
 
 
 def gaussian(x, y, sigma=1.0):
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     x, y = process_tensor(x, y)
     dist = np.linalg.norm(x - y, 2)
     return np.exp(-dist * dist / (2 * sigma * sigma))
@@ -48,8 +48,8 @@ def gaussian(x, y, sigma=1.0):
 
 def gaussian_tv(x, y, sigma=1.0):
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     x, y = process_tensor(x, y)
 
     dist = np.abs(x - y).sum() / 2.0
